LPS-71452 Move
